# packages/contracts/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LearnerMemoryCreate": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "memory_type": {
            "$ref": "#/components/schemas/LearnerMemoryType"
          }
        },
        "required": [
          "memory_type",
          "content"
        ],
        "title": "LearnerMemoryCreate",
        "type": "object"
      },
      "LearnerMemoryPatch": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "memory_type": {
            "$ref": "#/components/schemas/LearnerMemoryType"
          }
        },
        "required": [
          "memory_type",
          "content"
        ],
        "title": "LearnerMemoryPatch",
        "type": "object"
      },
      "LearnerMemoryResponse": {
        "properties": {
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "course_id": {
            "title": "Course Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "last_confirmed_at": {
            "format": "date-time",
            "title": "Last Confirmed At",
            "type": "string"
          },
          "memory_type": {
            "$ref": "#/components/schemas/LearnerMemoryType"
          },
          "source_kind": {
            "title": "Source Kind",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "course_id",
          "memory_type",
          "content",
          "confidence",
          "source_kind",
          "last_confirmed_at",
          "created_at",
          "updated_at"
        ],
        "title": "LearnerMemoryResponse",
        "type": "object"
      },
      "LearnerMemoryType": {
        "enum": [
          "preference",
          "confirmed_misconception",
          "learning_goal"
        ],
        "title": "LearnerMemoryType",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "期末复习智能体 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/courses/{course_id}/learner-memories": {
      "get": {
        "operationId": "list_learner_memories_api_v1_courses__course_id__learner_memories_get",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LearnerMemoryResponse"
                  },
                  "title": "Response List Learner Memories Api V1 Courses  Course Id  Learner Memories Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Learner Memories",
        "tags": [
          "learner-memories"
        ]
      },
      "post": {
        "operationId": "create_learner_memory_api_v1_courses__course_id__learner_memories_post",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LearnerMemoryCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearnerMemoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Learner Memory",
        "tags": [
          "learner-memories"
        ]
      }
    },
...
    "/api/v1/learner-memories/{memory_id}": {
      "delete": {
        "operationId": "delete_learner_memory_api_v1_learner_memories__memory_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "memory_id",
            "required": true,
            "schema": {
              "title": "Memory Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Learner Memory",
        "tags": [
          "learner-memories"
        ]
      },
      "put": {
        "operationId": "update_learner_memory_api_v1_learner_memories__memory_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "memory_id",
            "required": true,
            "schema": {
              "title": "Memory Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LearnerMemoryPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearnerMemoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Learner Memory",
        "tags": [
          "learner-memories"
        ]
      }
    },
...
  }
}